+++ Riot/Modules/Settings/SettingsViewController.m
@@ -137,6 +137,8 @@ enum
     LABS_USE_ROOM_MEMBERS_LAZY_LOADING_INDEX = 0,
     LABS_USE_JITSI_WIDGET_INDEX,
     LABS_CRYPTO_INDEX,
+    LABS_DM_KEY_VERIFICATION_INDEX,
+    LABS_CROSS_SIGNING_INDEX,
     LABS_COUNT
 };
 
@@ -2447,6 +2449,30 @@ SettingsIdentityServerCoordinatorBridgePresenterDelegate>
 
             cell = labelAndSwitchCell;
         }
+        else if (row == LABS_DM_KEY_VERIFICATION_INDEX)
+        {
+            MXKTableViewCellWithLabelAndSwitch* labelAndSwitchCell = [self getLabelAndSwitchCell:tableView forIndexPath:indexPath];
+            
+            labelAndSwitchCell.mxkLabel.text = NSLocalizedStringFromTable(@"settings_labs_dm_key_verification", @"Vector", nil);
+            labelAndSwitchCell.mxkSwitch.on = RiotSettings.shared.enableDMKeyVerification;
+            labelAndSwitchCell.mxkSwitch.onTintColor = ThemeService.shared.theme.tintColor;
+            
+            [labelAndSwitchCell.mxkSwitch addTarget:self action:@selector(toggleLabsDMKeyVerification:) forControlEvents:UIControlEventTouchUpInside];
+            
+            cell = labelAndSwitchCell;
+        }
+        else if (row == LABS_CROSS_SIGNING_INDEX)
+        {
+            MXKTableViewCellWithLabelAndSwitch* labelAndSwitchCell = [self getLabelAndSwitchCell:tableView forIndexPath:indexPath];
+            
+            labelAndSwitchCell.mxkLabel.text = NSLocalizedStringFromTable(@"settings_labs_cross_signing", @"Vector", nil);
+            labelAndSwitchCell.mxkSwitch.on = RiotSettings.shared.enableCrossSigning;
+            labelAndSwitchCell.mxkSwitch.onTintColor = ThemeService.shared.theme.tintColor;
+            
+            [labelAndSwitchCell.mxkSwitch addTarget:self action:@selector(toggleLabsCrossSigning:) forControlEvents:UIControlEventTouchUpInside];
+            
+            cell = labelAndSwitchCell;
+        }
     }
     else if (section == SETTINGS_SECTION_FLAIR_INDEX)
     {
@@ -3496,6 +3522,20 @@ SettingsIdentityServerCoordinatorBridgePresenterDelegate>
             }];
         }
     }
+}
+    
+- (void)toggleLabsDMKeyVerification:(id)sender
+{
+    UISwitch *switchButton = (UISwitch*)sender;
+    
+    RiotSettings.shared.enableDMKeyVerification = switchButton.isOn;
+}
+    
+- (void)toggleLabsCrossSigning:(id)sender
+{
+    UISwitch *switchButton = (UISwitch*)sender;
+    
+    RiotSettings.shared.enableCrossSigning = switchButton.isOn;
 }
 
 - (void)toggleBlacklistUnverifiedDevices:(id)sender